+++ CHANGELOG.rst
@@ -8,6 +8,8 @@ Changelog
 
 - Convert client.Request to base protocol
 - Rename proto to message
+- Extract server.Inspector
+
 
 0.2.2 (2015-04-10)
 ++++++++++++++++++